+++ src/tabs/synthesis/row.ts
@@ -1,5 +1,7 @@
 import { CheerioAPI, Cheerio, Element } from "cheerio";
+import copyrightedDictionaries from "../../constants/copyright.js";
 import Selectors from "../../constants/selectors.js";
+import { ParserOptions } from "../../options.js";
 
 interface WithMetadata<T> {
 	tags: string[];
@@ -11,14 +13,34 @@ interface Contents extends WithMetadata<string> {}
 
 export interface Row extends Contents {}
 
-export function parse($: CheerioAPI, row: Cheerio<Element>): Row {
-	const contents = getContents($, row);
+export function parse($: CheerioAPI, row: Cheerio<Element>, options: ParserOptions): Row | undefined {
+	const contents = getContents($, row, options);
+	if (contents === undefined) {
+		return undefined;
+	}
 
 	return { ...contents };
 }
 
-function getContents($: CheerioAPI, row: Cheerio<Element>): Contents {
+function getContents(
+	$: CheerioAPI,
+	row: Cheerio<Element>,
+	{ excludeCopyrighted }: ParserOptions,
+): Contents | undefined {
 	const section = row.children(Selectors.contentTabs.synthesis.body.row.contents.element);
+
+	const sources = section
+		.children(Selectors.contentTabs.synthesis.body.row.contents.sources)
+		.children()
+		.map((_, tag) => $(tag).text().trim())
+		.toArray();
+	if (excludeCopyrighted) {
+		const isCopyrighted = sources.some((source) => !copyrightedDictionaries.includes(source));
+		// If every source is copyrighted, reject the entry.
+		if (isCopyrighted) {
+			return undefined;
+		}
+	}
 
 	const tags = section
 		.children(Selectors.contentTabs.synthesis.body.row.contents.tags)
@@ -26,11 +48,6 @@ function getContents($: CheerioAPI, row: Cheerio<Element>): Contents {
 		.map((_, tag) => $(tag).text())
 		.toArray();
 	const text = section.children(Selectors.contentTabs.synthesis.body.row.contents.text).text().trim();
-	const sources = section
-		.children(Selectors.contentTabs.synthesis.body.row.contents.sources)
-		.children()
-		.map((_, tag) => $(tag).text().trim())
-		.toArray();
 
 	return { tags, value: text, sources };
 }