+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.1.0
 
 - Migrated package from Deno to Node.
+- Added option to exclude copyrighted dictionaries, set to `true` by default.